Replace core-image-minimal with core-image-base
Having core-image-minimal as a default first-on-the-list image
can be misleading when adding new boards to the build system
as this image allows only to boot-up minimal system without
any features that the board may support in terms of connectivity, etc.

core-image-full-cmdline, likewise, does not include "kernel-modules"
and can cause similar confusion.

Closes: https://git.ostc-eu.org/OSTC/OHOS/meta-ohos/-/issues/7
Signed-off-by: Robert Drab <robert.drab@huawei.com>