ohos-googletest: Integrate OHOS snapshot
ohos-googoletest is a fork of googletest with OHOS-specific extensions. Build
with make instead of OHOS specific and unusable BUILD.gn file. Unlike the
BUILD.gn file, build shared libraries for gtest and gtest_main. The remaining
gmock library is not yet needed by any test so it is not built.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>