ohos-googletest: switch to current ostc branch
All of the staging components are moving to a model where the master
branch follows upstream open harmony gitee repository and the ostc
branch is re-based on that master, and contains additional patches.

The zmk patch was applied at the "local" upstream (ostc branch), so drop
if from the recipe.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>