Build openharmony-image-base-tests for CI
We have notced that some of ACTS dependencies do not build correctly for
Avenger96. This went unnoticed because the CI system is building the
plain image, that does not include any test dependencies.

Fix that by building the test image instead of the plain image, for
targets based on Linux.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>