zmk: retain customized CC, CXX
ZMK has logic to set CC and CXX to the proper triplet name when the
configure script is invoked with different --build and --host. Normally
this works fine but when the configure script is also invoked with the
environment that picks a specific CC or CXX, such as by passing extra
flags through it, things can misbehave.

Vendor a patch from upstream until the next zmk release.

Fixes: https://git.ostc-eu.org/OSTC/OHOS/meta-ohos/-/issues/21
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>