systemd: Avoid musl warning
As part of our investigation and support, musl integration in the
project is fully supported (subject to the project's scope and QA
processes).

Upstream also carried out the relevant analysis and nothing 'known' is
broken: https://www.openwall.com/lists/musl/2019/11/08/6

That being said, we clean up the build logs of the respective warning.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>