systemd-conf: bring back dhcp for wired interfaces on Qemu
This file gets removed in the upstream poky recipe for qemuall due to
problems with it in the Yocto autobuilder QA.

On our side we need it in combination with spread so we bring it back
here. Next step would be to investigate why it fails on Yocto QA and get
it fixed upstream.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>