flavours: unify and symlink bblayers between different flavours
Our bblayers file diverged between the different flavours and e.g. the
meta-ohos-foundation was still present for zephyr and freertos.

We enable all layers on all flavours. The only overhead should be in
parsing the recipes. If something else breaks we consider this a bug
worth fixing.

The usage of symlinks should be safe here. Yocto buildhost will almost
always be Linux based, but it this should even work on Windows by now.
If we run into trouble with this we can always got back to three files.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>