systemd-conf: allow for runqemu and networkd compatibility
When we started to use the networkd unit for wired in qemu we also broke
the static IP configuration the runqemu script uses. The IP address, etc
are given as kernel commandline options for the static setup.

If we see this happening we can simply skip the networkd unit file for
DHCP setup as we know a static IP config was provided.

The key part of the imported unit file here is the KernelCommandLine=!ip
line in the match section to skip the unit if IP is present in the
commandline.

Closes: https://git.ostc-eu.org/OSTC/OHOS/meta-ohos/-/issues/29

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>