epiphany-app: Add support for running epiphany in application mode
This support is done configurable using the following variables:

EPIPHANY_APP             - Defines the app name (which needs to be the
			   directory name in FILESPATH/FILEEXTRAPATHS).
			   Currently defines it for HomeAssistant. If
			   this directory provides a
			   `service-override.conf`, the build system
			   will set it up as a systemd service
			   configuration for the main boot systemd
                           service.
EPIPHANY_URL             - The URL that the browser will load at
                           startup.
EPIPHANY_RDEPENDS        - Extra dependencies. If the specific
			   application needs extra dependencies at
                           runtime, they can be specifed here.
EPIPHANY_SERVICE_ENABLED - By default disabled (0). When enabled (1), it
			   will enable loading the epiphany app at boot.

The configuration defaults to setting up Epiphany browser in application
mode for a local instance of Home Assistant.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>