openharmony-image-common.inc: always include the gdbserver for debugging
By just installing this into all of our images we open the door for
in-field debugging. It needs to be invoked manually and we do not drag
in any dbg packages so the image size stays sane.

Combined with a remote debugging setup and copy of the images files and
symbols this can allow full debugging capabilities with GDB or an IDE
later on.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>