flavours: Use OSTC download and sstate-cache by default
OSTC now publishes a complete sstate-cache and download cache for
a set of builds and configurations. Using those caches can dramatically
speed up setting up the development environment by avoiding huge source
downloads (about 12GB at the time of this writing) and virtually all of
the compilation (close to an hour on a 6 core 8th gen Intel mobile CPU)
at the cost of miniscule sstate cache download (on the order of one
gigabyte).

The setting is added at the bottom of local.conf, where it is easier to
spot and disable, if desired.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>