.ostc-ci: Use shared .build job
This further reduces some copying by including build-generic.yaml from
the manifest repository. Bulk of the logic is reused now, with the
switch to a custom branch of the meta-ohos git repository being the sole
exception.

One small optimization, that is also applied to the manifest repository
is the override to drop image artifacts. Since those are not needed yet
we can save the cost of performing the uploads to GitLab object storage.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>