gitlab-ci.yml: restrict job creation based on files changed
CI process in the meta-ohos repository is rather heavy, with large
number of long-running build jobs. The pipeline system allows
restricting job startup to a set of rules, and to further restrict each
rule to trigger only when one of a set of given files is modified.

This should allow efficient modification pipeline that only touches the
documentation or the meta-layers, without needlessly running unrelated
testing.

This is documented at https://docs.gitlab.com/ee/ci/yaml/#ruleschanges

Closes: https://git.ostc-eu.org/OSTC/OHOS/meta-ohos/-/issues/34
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>