zephyr-blueprint-smarthome-base: substitute MAC address of a device
Multiple IoT devices in the Smart Home blueprint are to be
programmed with firmware built from the same sources. That
sources specify MAC address of the BLE interface of the device.
Multiple devices cannot work at once with the same MAC, so in
that case user should have possibility to specify the address
in build time.

Look for `MAC` variable - if set, use its value as MAC address,
substituting the default value in the source code of the firmware.

Signed-off-by: Wojciech Zmuda <wojciech.zmuda@huawei.com>
Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>