zephyr-blueprint-smarthome-sensors: substitute MAC address of a device
Multiple IoT devices in the Smart Home blueprint are to be
programmed with firmware built from the same sources. That
sources specify MAC address of the BLE interface of the device.
Multiple devices cannot work at once with the same MAC, so in
that case user should have possibility to specify the address
in build time.

Variable SMART_HOME_SENSORS_MAC can be provided to Bitbake.
If detected, its value is used as MAC address for devices
using zephyr-blueprint-smarthome-sensors image.

Signed-off-by: Wojciech Zmuda <wojciech.zmuda@huawei.com>
Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>