smart-home-homeassistant-plugin: Add recipe
The recipe installs Home Assistant custom component for communication
with IoT sensors and light switch devices running All ScenariOS firmware
for the Smart Home blueprint.

The recipe adds entries to HA configuration.yaml file. The entries are
generated with MAC addresses of the expected devices, that are part of
the blueprint setup. They are defined in the relevant blueprint config
file and can be overriden in local.conf.

Signed-off-by: Wojciech Zmuda <wojciech.zmuda@huawei.com>
Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>