homeassistant-config.inc: Provide an inc file for the HA configuration
This inc file is useful so multiple recipes can synchronise on the same
path for the main Home Assistant configuration. For example, a recipe
bringing a custom Home Assistant configuration would use this as the
source of truth for the directory where this configuration needs to
reside. Home Assistant will use the same configuration path at
runtime.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>