Add machine and flavour aware recipes from manifest
Currently machines and flavours are defined in the meta-ohos repository
but their CI pipelines are defined in the manifest repository. Move the
definitions of the various build jobs here to be included from the
manifest repository internally.

This way the public interface is unchanged but patches such as the
recent Raspberry Pi 4 support can enable a new MACHINE configuration and
a new CI job all in one go, to the extent that in-progress CI will pick
it up.

Closes: https://git.ostc-eu.org/OSTC/OHOS/meta-ohos/-/issues/44
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>