.ostc-ci: include local machines-and-flavours.yaml
Include the machines-and-flavours file explicitly, from the local
repository.  It is already included via the build-generic.yaml
project-include above, but including it again here allows GitLab to pick
up the changes to the definition straight away. The include done via the
manifest repository is using the default branch, so it cannot pick up
changes to this file as it is being modified by incoming branches.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>