wpantund: Fix build break with missing strlcpy symbol
The wpantund project has some m4 macro to check for strlcpy and strlcat
being available. We are hitting a corner case here (maybe due to musl)
where strlcpy is detected but the system still does not build.

Adding a patch to the project to fix this case which was also submitted
upstream and waits for review.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>