allscenarios-linux-blueprint-gateway.conf: Initial distro and image
Basic distribution for the Gateway Blueprint as well as a gateway  image
that pulls in all networking packages we are using for this first
iteration.

Two blueprint specific overrides are added as well. We make sure hostapd
service gets started by default for a good ootb experience. Secondly we
enable the iptables package config flag for systemd-netword to have the
basic NAT functionality enabled.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>