hostapd: Gateway blueprint WiFi AP configuration
Basic WiFi AP configuration for nl80211 driver with WPA-PSK to avoid
opening up your networking accidentally to everyone.

The appended systemd service file fixes an issue with the upstream one
on the command-line argument ordering. The conf file has to be last or
the service does not start for me.

Fixes: https://git.ostc-eu.org/OSTC/planning/core-os/-/issues/83

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>