Update binutils to 2.36.1
To make use of some features introduced in the clang 12 based toolchain,
such as support for DWARF-5 format debug info, support in binutils
(in particular as and ld) is needed.

Bring binutils up to 2.36.1 (in sync with patches in poky master) to have
a binutils version matching what some newer toolchain features need.

The recipe is in sync with the recipe in poky revision
e7ce28c8b060fc4fc84fe328275bbca87d17a3f0

Signed-off-by: Bernhard Rosenkränzer <bernhard.rosenkraenzer.ext@huawei.com>