matter: add initial recipe
Recipe for the Matter networking protocol (formerly known as CHIP). This
project has more the form of a vendor SDK with support for many bare
metal SDK's as well as various operating systems.

This recipe does focus on the Linux build and configures the project to
disable all extra third_party git submodules (we only use 6 out of 48).
Matter uses the GN meta build system, therefore using gn.bbclass as a base
for the recipe. gn.bbclass requires patching BUILDCONFIG.gn file to set default
toolchain and add compiler/linker flags to the yocto-provided ones.

Mbedtls version integrated in Matter doesn't compile cleanly with GCC
v11. It is not required by currently used Linux target but in case it's needed
version provided by the Yocto build system shall be used instead of this
integrated one and therefore disabling building of the mbedtls completely.

Signed-off-by: Robert Drab <robert.drab@huawei.com>
Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>