allscenarios-linux.conf: hardening, remove NFS
We do not use NFS in the image, nor in any of the blueprints. An unused service
is a potential security issue, so we remove it.

In addition to expected DISTRO_FEATURES/IMAGE_FEATURES, we need to remove a
dependency to the packagegroup-core-device-devel that includes nfs. This
package is included in poky.conf for all qemu builds, but is not needed and
has been removed from Poky [1] in August 2020.

[1] https://git.yoctoproject.org/cgit/cgit.cgi/poky/commit/?id=d707fa30f8a24d1e50831846330757254f245791

Signed-off-by: Marta Rybczynska <marta.rybczynska@huawei.com>