x-wic.inc: Use a copy of the rootfs in ROOTB
The root b partition was initially set up using the assumption that a
wic part configuration without fstype would leave the partition empty
(no filesystem). Checking a bit deeper the wic implementation, that
never worked as expected as the tool defaults to a vfat partition[1] (when
no fstype is provided). Also, squashfs doesn't support an empty
filesystem.

Given the above, we switch the second rootfs (B) to using the same
content as the default one (A). This has an added advantage that we can
test switching the root filesystem A-B without any additional
operations.

[1] https://git.yoctoproject.org/cgit/cgit.cgi/poky/tree/scripts/lib/wic/ksparser.py?h=dunfell#n158

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>