bootimg-pcbios-label.py: Workaround the label limitation of bootimg-pcbios
The bootimg-pcbios wic plugin in dunfell hardcodes the filesystem label
without taking into consideration the --label argument. This was
improved later on but it never landed in dunfell:

commit a79b5f0f2139bbd77df31704944630c8a5a0b562
Author: jbouchard <jeanbouch418@gmail.com>
Date:   Sun Jun 27 20:28:37 2021 -0400
Use the label provided when formating a dos partition

We workaround this by creating a version of this plugin that has this
change included and we will be using this new plugin until the
improvement (if at all) lands in dunfell.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>