ohos-image.bbclass: Revert dropbear hostkey dir configuration
When read-only rootfs is enabled - it is enabled default for the
reference images - the core bbclass rootfs-postcommands.bbclass
configures dropbear to use a volatile location (a tmpfs location) for
the hostkey because the default location, /etc/dropbear is read-only.
With the advent of writables.bbclass, the OS now provides the default
path /etc/dropbear as a writable (a persistent location bind-mounted
from the sysdata partition). This makes now /etc/dropbear both
read-write and also persistent over reboots. This change reverts the
rootfs-postcommands.bbclass dropbear configuration to use a volatile
location as opposed to the default one (/etc/dropbear).

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>