x-wic.inc: Rename partition labels
Our initial implementation defines the partitions labels as it follows:
x-boot
x-dev-data
x-sys-data
x-app-data

This was proven to be problematic in many ways.

First of all, this needs to be treated with care as the associated
systemd mount unit would need to escape the dashes. For example,
"x-dev-data" would need to have a mount unit associated escaping this to
"x\x2ddev\x2ddata".

Secondly, the yocto version we are using (dunfell) has a bug where
backslashes are not supported in SRC_URI[1]. This would require a
workaround when using mount points with an escaped filename.

The easiest way forward would be to just not use dashes for the labels.
In this way, we maintain consistentcy for the mountpoints (/mnt/<label>)
without having to deal with all the above.

[1] https://bugzilla.yoctoproject.org/show_bug.cgi?id=8161

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>