sysota: Integrate git snapshot
SystemOTA or sysota is a robust, unattended update system for Linux
Gateways, suitable for home automation gateways and other devices where
reliability is paramount.

This base recipe builds fine and works in Qemu. On a Raspberry Pi system
it needs to be paired with an appropriate RAUC configuration file and a
matching SystemOTA configuration file. Those will be provided with
specialized packages as the details are closely related to WIC images.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>