xtc-acts: Don't break the build if -Werror=array-bounds is enabled
This is a workaround for a legitimate xtc-acts bug modern compilers complain
about when extra warnings for potential security problems are enabled.

While this is not the proper fix and this should be fixed properly at some
point (reading 16 bytes from a 12-character string is not a good idea...),
this is enough to keep the build working until someone has time to fix it
properly, and it doesn't make things worse (before, this just went
unnoticed because the warning wasn't enabled).

Signed-off-by: Bernhard Rosenkränzer <bernhard.rosenkraenzer.ext@huawei.com>