binutils: Add backport to fix running into to many open files on thin archives
This problem showed when building nodejs and resulted in an error of
malformed archives:

libv8_initializers.a: error adding symbols: malformed archive

There was a corner case where the fd's have not been closed on thin
archives and nodejs was heavy enough to be linked to show this problem.

Upstream report and analysis can be seen here:
https://sourceware.org/bugzilla/show_bug.cgi?id=28138

Fix is already upstream and just a backport for 2.37 added here.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>