.ostc-ci: route linux blueprints to "heavy" runners
With the recent improvements to the CI infrastructure we are seeing
relatively fast build times for most of our jobs. The exception to this
rule are the relatively-larger blueprint build jobs.

Assign the set of tags: "bitbake-friendly", large-disk (legacy) and the
new "heavy" tag. The tag will route the job to a runner with
significantly more resources. This shoudl help even out the total build
time.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>