.ostc-ci: switch dual linux builds to parallel matrix
With the .build-linux-matrix job now available from the manifest
repository, we can now switch linux-seco-intel-b68,
linux-seco-imx8mm-c61 and linux-raspberrypi4-64 to build the base and
extras images separately, as distinct jobs.

In the end the exact same thing happens but we have more parallelism
since the jobs can be picked up by separate workers.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>