allscenarios-image*-tests: Add package testing to the test images
This is the first step in initiating automated package testing for
images. There will need to be some work on the CI side and some work on
the entire process but this is a first pass at getting the images to the
point where they have ptest on them.

This commit also adds an ext4 filesystem so we can test a writable
image. The wic image is still r/o but there is some benefit to having
tests run on a r/w fs.

Signed-off-by: Eilís Ní Fhlannagáin <elizabeth.flanagan@huawei.com>