.ostc-ci: use $OHOS_RUNNER_TAG on blueprint jobs
With the introduction of $OHOS_RUNNER_TAG on the manifest repository, we
can now reliably add a custom tag to all bitbake builds that do not
customize the tag list.

Use the new tag on the two jobs which provide a custom list to be able
to redirect them to a special runner as well.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>