rauc: use known insecure keys by default
Unless overridden in another layer or in local.conf, rauc will bundle
and use a know public key for verifying updates. This allows us to
to have some set of defaults that work out of the box and allow testing
the OTA stack and producing signed update bundles coming out of the CI
system.

The key is insecure and will expire in a year. This is intentional. Over
time we may switch to a reference key that is managed and secured or
re-generate another key for another year.

The insecure key has a fixed name and that name is used to trigger a
bitbake warning. OTA stack will learn to detect that key and similarly
warn users, or perform equally appropriate operations, in order to
reduce the risk of someone accidentally using this key in production.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>