qemu: Define libnfs PACKAGECONFIG
The upstream qemu recipe uses host's pkg-config files as a solution to
detecting host's SDL[1]. This has a side effect of using other host
libraries that are later queried by the configure script. This can get
into a situation when the host provides libnfs (for example) and because
later this dependency is not in place anymore, qemu will fail at
runtime.

This change adds a PACKAGECONFIG definition for libnfs that is disabled
by default, in turn disabling by default pkgconfig autodetection in
configure.

[1] https://git.yoctoproject.org/cgit/cgit.cgi/poky/tree/meta/recipes-devtools/qemu/qemu.inc?h=dunfell#n165

Fixes: https://git.ostc-eu.org/OSTC/OHOS/meta-ohos/-/issues/141
Upstream-status: Submitted [https://lists.openembedded.org/g/openembedded-core/message/156694]
Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>