ostc-ci: Workaround for inter-repo dependency on image builds
For the images renames to fully pass CI, we need to have a change in the
manifest repo. As these two MR would block each other before we could
merge them, we override the image names here, land the rename, do the
actual change in the manifest repo and finally clean this up again.

To be reverted when this change lands in the manifest repository.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>