.ostc-ci: use build-wic-image for qemu
Our QEMU builds were using the .build-image job, which archives all of
the deploy/image directory. This directory started to inflate recently,
in ways I did not explore.

Given that this breaks CI and is split across repositories the
.build-wic-image job was pushed into the manifest repository. Now with
the new job available we can switch the x86 and x86_64 QEMU image builds
to use the new .build-wic-image job, which saves a subset of the
artifacts that CI actually consumes downstream.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>