oniro-linux-blueprint-gateway: Disable dnsmasq service by default
In our core distributions we disable the dnsmaq packageconfig for
networkmanager. Re-enable it here as we need it for the Soft AP use
case.
If dnsmasq runs as active systemd service by default it conflicts with
network-manager which calls it directly. For this blueprint we want to
leave the control in networkmanagers realm.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>