manifests: Import XML files from manifests repository
The repo manifests are currently in a separate repository. This became a
burden in maintenence as meta-asos (oniro) has hard dependencies on it
(for example bblayers configuration sample). This moves the manifests in
the main oniro repository so we can handle repo manifests and build
metadata changes atomically.

The manifests repository is to be deprecated.

The import of the manifest XML files is clean with the exeption of
flattening the projects in sources. In other words, a project that used
to be checked out in /sources/foo is not checked out in /foo.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>