Rename meta-ohos-staging to meta-oniro-staging
This change is in line with the project's rebranding.

The collection is also changed a bit to drop the "meta-" prefix. There
is no clear standardization on the collection name. Some use an ID, some
use ID-layer, some use meta-ID etc. We have decided to use the the layer
name without the "meta-" prefix  because this is an ID, not a layer per
se.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>