.oniro-ci,docs: merge CI from former manifest repo
Most things are moved unchanged. The relatively few changes are to the
documentation, as the split between the meta-ohos and manifest is gone,
as well as to the actual CI definitions for the workspace directory,
since the sources/ sub-directory is gone, and there was some
backwards-compatibility code that no longer makes sense to maintain.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>