sysota: move to recipes-core
The recipes-ota directory was an earlier attempt to package ota as well
as the set of Go dependencies that turned out to be impossible to
package at the time, due to go-mod.bbclass changes they would require.

Since it's just one package, move it to recipes-core, next to rauc
overrides.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>