rauc: ensure RAUC state directory exists
RAUC uses a state file `status.raucs`. In the Oniro images this file is
prescribed to be in `/run/mount/sysdata/common/`. Exploratory testing
has found that RAUC cannot create the leading directory automatically
and fails to handle an update transaction.

Resolve this by adding a systemd configuration override file which
creates the directory using ExecStartPre=

Closes: https://git.ostc-eu.org/distro/oniro/-/issues/158
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>