oniro-linux.conf: Fix wic deploy of overlay_map.dtb
overlay_map.dtb is to be deployed in the overlays subdirectory but
dunfell only decides on what to deploy there based on extension. An
exception for overlay_map.dtb was added in later versions but was never
backported as it only makes sense on newer kernel which dunfell doesn't
support. We fix this manually in IMAGE_BOOT_FILES.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>