.gitlab-ci.yml: fix testing changes to the manifest
The manifest is special, because when it changes the oniro repository is
cloned three times inside the CI process:

- first time by GitLab, as a part of the normal CI checkout process
- second time by repo init, as a part of the .workspace job
- last time by repo sync, as a part of the same job

When the manifest changes, the CI system should be testing the incoming
change. Crucially the second step needs to be adjusted, to take the
manifest from a non-default location. This was handled before the
manifest repository was merged into the oniro repository. During that
merge the critical step was lost.

Adjust the .bitbake-workspace job to inject variables that pick the
correct manifest repository and branch. Leave a comment for the untested
rule about CI for tags.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>