sysota: persist state directory
SystemOTA needs one bit of information, namely, if the system is in
try-boot mode. This information may not be available from the boot
system at all times, and is required for correct operation.

Persist the entire /var/lib/sysota directory, as SystemOTA creates
temporary files and renames over the desired filename for atomicity.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>