Revert "manifests: default.xml: bump meta-seco-imx"
This reverts commit 07f09c1a9f7284b9974fdce7d4aee6ca175ef3c8.

This introduced a build error on linux-seco for C61. Due to this merge
happening while the GitLab transition was in progress, the CI didn't
pick it up and it was merged without a CI check. We are currently
investigating the build issue and while doing it, we revert the
offending commit.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>