otbr-configuration: Enhance configuration to setup Commissioner Credential
This was a key part missing to make our own configuration work with an
external commissioner, like the Android app.

Also removing the space in the network name for now. Not sure if this
has been part of the problems I have seen, but none of the example has a
space in the name so better avoid it until fully debugged.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>