networkmanager: Change PACKAGECONFIG to a direct assignment
Having the PACKAGECONFIG_remove in an bbappend file makes it really hard
to consume this layer and bring the specific config back with an _append
(happens for me with the gateway blueprint.

Instead we are now making a direct assignment, with he values removed
taken into account, from our distro as policy. This allows other distros
or layers consuming meta-oniro-core to change it as needed.

A new oniro-packageconfig.inc include file is created in which we will
consolidate other PACKAGECONFIG changes in the next commit.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>