ci: don't run lava test if required var CI_LAVA_TOKEN not set
CI_LAVA_TOKEN is required for submitting job to LAVA server. It should
be set in instance level gitlab CI/CD variables for testing forking dev
flow based MRs.

With `when: never` rule, test job will not be added to pipeline if the
above var not set. If the condition not set in test job that extends
`.lava-test`, `.lava-test` will ask user to do so and then exit with
unsuccessful job status. There is no easy way to exit early with
successful status yet. See
https://gitlab.com/gitlab-org/gitlab/-/issues/329217#note_572963049

Adding the above condition in `.lava-test` has no impact on test job
that extends the `.lava-test`.

Signed-off-by: Chase Qi <chase.qi@linaro.org>