binutils: Drop version 2.37
oe-core now comes with this version already: 2.37. 2.37 in oniro was an
upgrade for dunfell but now that we move to newer layers, we can drop it
and use the latest in upstream.

We can safely drop
0017-bfd-Close-the-file-descriptor-if-there-is-no-archive-fd.patch as it
was included in oe-core.

We also split the hardening change done in
95bd0062b2293bf0b8af05060e958923574933ff in a separate bbappend. This is
to be reviewed for upstream inclusion.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>