ot-br-posix: Move writables usage from staging to meta-oniro-core bbappend
This is a follow up fix for e791c29b819a629930bbde6fd7e23bdd9277f31f

The writables part is correct, but it can not live in the
meta-oniro-staging layer. This layer is in preparation of upstreaming
and the writables class we have in core is not at this stage yet.

Instead we split out the writables part into a bbappend to live in
meta-oniro-core. This solved an implicit layer dependcy from staging to
core as well which we saw with the meta-openharmony build.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>