bluez: fix CVE-2021-0129
Improper access control in BlueZ may allow an authenticated user to
potentially enable information disclosure via adjacent access.

This issue can be fixed in the kernel, in BlueZ or both. This patch
fixes it on the BlueZ side, so that the configuration no longer
depends on the kernel fix.

https://nvd.nist.gov/vuln/detail/CVE-2021-0129

Upstream-Status: Submitted [https://lists.openembedded.org/g/openembedded-core/message/159682]

Signed-off-by: Marta Rybczynska <marta.rybczynska@huawei.com>