sysota: update to v0.1.0
This version of SysOTA is capable of network updates and rollbacks and
seems feature complete. It has some missing tests and validation of
non-essential properties.

The package is now both stateful and configurable. Configuration is
persisted in /etc/sysota. This allows a device to remember the location
and settings of the update server.

The new sysotactl executable allows configuring the update service and
performing updates. The executable is a symbolic link to sysota-mux. The
sysotad executable is also converted to a symbolic link to the mux
binary.

The default configuration files name the device and model as Oniro
Project Reference Device. Raspberry Pi is named more precisely.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>