Change license for all 1st party patches to MIT
While Apache 2.0 license is the recommended option when choosing an open
source permissive (i.e. non-copyleft) license, such license is
incompatible with some other open source licenses (namely, GPL-2.0) so
it is not the best suited choice for patch files, since patch files are
naturally intended to modify third party code, which at times may be
subject to incompatible licenses.

Instead of checking license compatibility on a case by case basis (which
not only is time consuming, but it also a hardly maintainable/scalable
process and, most important, error-prone), a more practical approach
would be to adopt an open source license with the broadest compatibility
possible, namely, the MIT license.

Signed-off-by: Alberto Pianon <pianon@array.eu>