wic: add x-gpt-efi-disk workspace file
This file is derived from QEMU+EFI file we already have while having
wider scope of supporting both virtual and physical machines alike.

The partition table is GPT and the boot system is based on EFI. The
EFI_PROVIDER variable can be used to select either grub or systemd-boot.

Over time the qemu-specific workspace files will be deprecated and
removed. This step makes the transition easier to process.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>