.oniro-ci: drop the compliance tag from REUSE check
Historically the REUSE check was routed to a dedicated but slow machine
that guaranteed that no CI would be stuck waiting on the miniscule
compliance job.

With the Eclipse migration and CI/CD being slightly out of our control,
it's just more convenient to drip this tag rather than request
configuration change in the instance.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>