rauc: adjust compatible string for Raspberry Pi 4
In retrospective I think I've misunderstood the compatible string to
mean "what is the image we expect" vs "what is the hardware we expect".
By setting the compatible string to describe the hardware, we get the
ability to restrict images for genuinely incompatible hardware without
locking down ability to move to other software images.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>